ISSUE #3466 - use self intercepting unhandled error component
