+++ frontend/src/v5/ui/components/shared/modals/templates/deleteModal/deleteModal.component.tsx
@@ -28,9 +28,8 @@ import {
 	Instruction,
 } from '@/v5/ui/components/shared/modals/modals.styles';
 import { CircledIcon } from '@controls/circledIcon';
-import { useErrorInterceptor } from '@controls/errorMessage/useErrorInterceptor';
 import { useForm } from 'react-hook-form';
-import { UnhandledError } from './deleteModal.styles';
+import { UnhandledErrorInterceptor } from './deleteModal.styles';
 
 interface IDeleteModal {
 	onClickClose?: () => void,
@@ -91,7 +90,7 @@ export const DeleteModal = ({ onClickConfirm, onClickClose, name, message, confi
 						/>
 					</RetypeCheck>
 				)}
-				<UnhandledError error={useErrorInterceptor()} />
+				<UnhandledErrorInterceptor />
 			</Message>
 			<Actions>
 				<Button onClick={onClickClose} variant="contained" color="primary">
