+++ frontend/src/v5/ui/controls/errorMessage/networkError/networkError.component.tsx
@@ -14,12 +14,15 @@
  *  You should have received a copy of the GNU Affero General Public License
  *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
+import { FormattedMessage } from 'react-intl';
+import { ErrorMessage } from '../errorMessage.component';
 
-import styled from 'styled-components';
-import {
-	UnhandledErrorInterceptor as UnhandledErrorInterceptorBase,
-} from '@controls/errorMessage/UnhandledErrorInterceptor/UnhandledErrorInterceptor.component';
+type NetworkErrorProps = {
+	className?: string;
+};
 
-export const UnhandledErrorInterceptor = styled(UnhandledErrorInterceptorBase)`
-	justify-content: center;
-`;
+export const NetworkError = ({ className }: NetworkErrorProps) => (
+	<ErrorMessage className={className}>
+		<FormattedMessage id="errorMessage.networkError" defaultMessage="Network Error" />
+	</ErrorMessage>
+);
