+++ frontend/src/v5/ui/controls/errorMessage/unhandledError/unhandledError.component.tsx
@@ -15,10 +15,10 @@
  *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 import { isNetworkError } from '@/v5/validation/errors.helpers';
-import { NetworkError } from './networkError/networkError.component';
+import { NetworkError } from '../networkError/networkError.component';
 import { UnexpectedError } from '../unexpectedError/unexpectedError.component';
 
-type UnhandledErrorProps = {
+export type UnhandledErrorProps = {
 	error: any;
 	expectedErrorValidators?: Array<(err) => boolean>;
 	className?: string;
