+++ frontend/src/v5/ui/controls/errorMessage/unhandledErrorInterceptor/unhandledErrorInterceptor.component.tsx
@@ -14,15 +14,12 @@
  *  You should have received a copy of the GNU Affero General Public License
  *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-import { FormattedMessage } from 'react-intl';
-import { ErrorMessage } from '../../errorMessage.component';
+import { UnhandledError, UnhandledErrorProps } from '../unhandledError/unhandledError.component';
+import { useErrorInterceptor } from '../useErrorInterceptor';
 
-type NetworkErrorProps = {
-	className?: string;
+type UnhandledErrorInterceptorProps = Omit<UnhandledErrorProps, 'error'>;
+
+export const UnhandledErrorInterceptor = (props: UnhandledErrorInterceptorProps) => {
+	const error = useErrorInterceptor();
+	return (<UnhandledError {...props} error={error} />);
 };
-
-export const NetworkError = ({ className }: NetworkErrorProps) => (
-	<ErrorMessage className={className}>
-		<FormattedMessage id="errorMessage.networkError" defaultMessage="Network Error" />
-	</ErrorMessage>
-);
