+++ frontend/src/v5/ui/routes/dashboard/projects/projectsList/createProjectModal/createProjectModal.component.tsx
@@ -26,8 +26,9 @@ import { FormSelect } from '@controls/formSelect/formSelect.component';
 import { MenuItem } from '@mui/material';
 import { TeamspacesHooksSelectors } from '@/v5/services/selectorsHooks/teamspacesSelectors.hooks';
 import { projectAlreadyExists } from '@/v5/validation/errors.helpers';
-import { UnhandledError } from '@controls/errorMessage/unhandledError/unhandledError.component';
-import { useErrorInterceptor } from '@controls/errorMessage/useErrorInterceptor';
+import {
+	UnhandledErrorInterceptor,
+} from '@controls/errorMessage/unhandledErrorInterceptor/unhandledErrorInterceptor.component';
 
 interface ICreateProject {
 	open: boolean;
@@ -124,7 +125,7 @@ export const CreateProjectForm = ({ open, onClickClose }: ICreateProject) => {
 				control={control}
 				formError={errors.projectName}
 			/>
-			<UnhandledError expectedErrorValidators={[projectAlreadyExists]} error={useErrorInterceptor()} />
+			<UnhandledErrorInterceptor expectedErrorValidators={[projectAlreadyExists]} />
 		</FormModal>
 	);
 };
