+++ frontend/src/v5/ui/routes/login/login.styles.tsx
@@ -17,7 +17,9 @@
 
 import styled from 'styled-components';
 import { Typography } from '@controls/typography';
-import { UnhandledError as UnhandledErrorBase } from '@controls/errorMessage/unhandledError/unhandledError.component';
+import {
+	UnhandledErrorInterceptor as UnhandledErrorInterceptorBase,
+} from '@controls/errorMessage/unhandledErrorInterceptor/unhandledErrorInterceptor.component';
 
 export const OtherOptions = styled.div`
 	display: flex;
@@ -41,6 +43,6 @@ export const ForgotPasswordPrompt = styled(StyledTypography)`
 	margin-left: auto;
 `;
 
-export const UnhandledError = styled(UnhandledErrorBase)`
+export const UnhandledErrorInterceptor = styled(UnhandledErrorInterceptorBase)`
 	margin-top: 5px;
 `;
