+++ frontend/src/v5/ui/routes/userSignup/userSignupForm/userSignupFormStep/userSignupFormStepTermsAndSubmit/userSignupFormStepTermsAndSubmit.component.tsx
@@ -23,9 +23,11 @@ import { UserSignupSchemaTermsAndSubmit } from '@/v5/validation/userSchemes/user
 import { clientConfigService } from '@/v4/services/clientConfig';
 import ReCAPTCHA from 'react-google-recaptcha';
 import { pick, defaults, isMatch } from 'lodash';
-import { useErrorInterceptor } from '@controls/errorMessage/useErrorInterceptor';
 import SignupIcon from '@assets/icons/outlined/add_user-outlined.svg';
-import { UnhandledError } from '@controls/errorMessage/unhandledError/unhandledError.component';
+import {
+	UnhandledErrorInterceptor,
+} from '@controls/errorMessage/unhandledErrorInterceptor/unhandledErrorInterceptor.component';
+
 import { emailAlreadyExists, usernameAlreadyExists } from '@/v5/validation/errors.helpers';
 import {
 	CreateAccountButton,
@@ -176,10 +178,7 @@ export const UserSignupFormStepTermsAndSubmit = ({
 					/>
 				)}
 			</TermsContainer>
-			<UnhandledError
-				expectedErrorValidators={[emailAlreadyExists, usernameAlreadyExists]}
-				error={useErrorInterceptor()}
-			/>
+			<UnhandledErrorInterceptor expectedErrorValidators={[emailAlreadyExists, usernameAlreadyExists]} />
 			<CreateAccountButton
 				isPending={formIsSubmitting || captchaIsPending}
 				startIcon={<SignupIcon />}
